use established Node-style callback conventions
In Node, the convention for asynchronous operations is to accept a
function as their last argument with the following signature:

    (error: any) => void
    (error: undefined, data: any) => void

In the event of an error, then the first argument passed to the callback
will be an error object (usually an actual instance of Error). If there
was no error, then the first argument will be falsey (undefined/null
usually) and the second argument will contain the data.